src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.1
asyncpg==0.29.0
# starlette[full]
transformers
torch
timm
open_clip_torch
scikit-learn
opencv-python-headless
Pillow
numpy
hdbscan
piexif
python-docx
aiofiles
aioshutil
matplotlib
pyproj
